requirements.txt
python-dotenv==0.21.0
pandas==1.3.3
numpy==1.21.2
psycopg2-binary==2.9.4
tqdm==4.62.3
scikit-learn==0.24.2
plotly==5.3.1
